+++ src/manifolds/Sphere.jl
@@ -6,7 +6,7 @@ An abstract type to represent a unit sphere that is represented isometrically in
 abstract type AbstractSphere{𝔽} <: AbstractDecoratorManifold{𝔽} end
 
 @doc raw"""
-    Sphere{T,𝔽} <: AbstractSphere{𝔽}
+    Sphere{𝔽, T} <: AbstractSphere{𝔽}
 
 The (unit) sphere manifold ``𝕊^{n}`` is the set of all unit norm vectors in ``𝔽^{n+1}``.
 The sphere is represented in the embedding, i.e.
@@ -45,16 +45,16 @@ and the [`zero_vector`](@ref zero_vector(::Euclidean, ::Any...)) are inherited f
 Generate the (real-valued) sphere ``𝕊^{n} ⊂ ℝ^{n+1}``, where `field` can also be used to
 generate the complex- and quaternionic-valued sphere.
 """
-struct Sphere{T, 𝔽} <: AbstractSphere{𝔽}
+struct Sphere{𝔽, T} <: AbstractSphere{𝔽}
     size::T
 end
 function Sphere(n::Int, field::AbstractNumbers = ℝ; parameter::Symbol = :type)
     size = wrap_type_parameter(parameter, (n,))
-    return Sphere{typeof(size), field}(size)
+    return Sphere{field, typeof(size)}(size)
 end
 
 @doc raw"""
-    ArraySphere{T<:Tuple,𝔽} <: AbstractSphere{𝔽}
+    ArraySphere{𝔽, T} <: AbstractSphere{𝔽}
 
 The (unit) sphere manifold ``𝕊^{n₁,n₂,...,nᵢ}`` is the set of all unit (Frobenius) norm elements of
 ``𝔽^{n₁,n₂,...,nᵢ}``, where ``𝔽\in\{ℝ,ℂ,ℍ\}. The generalized sphere is
@@ -65,7 +65,7 @@ tensors of unit norm. The set formally reads
 𝕊^{n_1, n_2, …, n_i} := \bigl\{ p \in 𝔽^{n_1, n_2, …, n_i}\ \big|\ \lVert p \rVert = 1 \bigr\}
 ````
 
-where ``𝔽\in\{ℝ,ℂ,ℍ\}``. Setting ``i=1`` and ``𝔽=ℝ``  this  simplifies to unit vectors in ``ℝ^n``, see
+where ``𝔽∈\{ℝ,ℂ,ℍ\}``. Setting ``i=1`` and ``𝔽=ℝ``  this  simplifies to unit vectors in ``ℝ^n``, see
 [`Sphere`](@ref) for this special case. Note that compared to this classical case,
 the argument for the generalized case here is given by the dimension of the embedding.
 This means that `Sphere(2)` and `ArraySphere(3)` are the same manifold.
@@ -76,10 +76,10 @@ The tangent space at point ``p`` is given by
 T_p 𝕊^{n_1, n_2, …, n_i} := \bigl\{ X ∈ 𝔽^{n_1, n_2, …, n_i}\ |\ \Re(⟨p,X⟩) = 0 \bigr \},
 ````
 
-where ``𝔽\in\{ℝ,ℂ,ℍ\}`` and ``⟨⋅,⋅⟩`` denotes the (Frobenius) inner product in the
+where ``𝔽∈\{ℝ,ℂ,ℍ\}`` and ``⟨⋅,⋅⟩`` denotes the (Frobenius) inner product in the
 embedding ``𝔽^{n_1, n_2, …, n_i}``.
 
-This manifold is modeled as an embedded manifold to the [`Euclidean`](@ref), i.e.
+This manifold is modelled as an embedded manifold to the [`Euclidean`](@ref), i.e.
 several functions like the [`inner`](@ref inner(::Euclidean, ::Any...)) product and the
 [`zero_vector`](@ref zero_vector(::Euclidean, ::Any...)) are inherited from the embedding.
 
@@ -89,7 +89,7 @@ several functions like the [`inner`](@ref inner(::Euclidean, ::Any...)) product 
 
 Generate sphere in ``𝔽^{n_1, n_2, …, n_i}``, where ``𝔽`` defaults to the real-valued case ``ℝ``.
 """
-struct ArraySphere{T, 𝔽} <: AbstractSphere{𝔽}
+struct ArraySphere{𝔽, T} <: AbstractSphere{𝔽}
     size::T
 end
 function ArraySphere(
@@ -98,7 +98,7 @@ function ArraySphere(
         parameter::Symbol = :type,
     ) where {I}
     size = wrap_type_parameter(parameter, n)
-    return ArraySphere{typeof(size), field}(size)
+    return ArraySphere{field, typeof(size)}(size)
 end
 
 """
@@ -201,7 +201,7 @@ function exp_fused!(M::AbstractSphere, q, p, X, t::Number)
     return q
 end
 
-function get_basis_diagonalizing(M::Sphere{<:Any, ℝ}, p, B::DiagonalizingOrthonormalBasis{ℝ})
+function get_basis_diagonalizing(M::Sphere{ℝ}, p, B::DiagonalizingOrthonormalBasis{ℝ})
     n = get_parameter(M.size)[1]
     A = zeros(n + 1, n + 1)
     A[1, :] = transpose(p)
@@ -247,7 +247,7 @@ end
 function get_embedding(M::AbstractSphere{𝔽}) where {𝔽}
     return Euclidean(representation_size(M)...; field = 𝔽)
 end
-function get_embedding(M::Sphere{<:Tuple, 𝔽}) where {𝔽}
+function get_embedding(M::Sphere{𝔽, <:Tuple}) where {𝔽}
     return Euclidean(representation_size(M)...; field = 𝔽, parameter = :field)
 end
 
@@ -337,12 +337,12 @@ return the local representation of the metric in a [`DefaultOrthonormalBasis`](@
 the diagonal matrix of size ``n×n`` with ones on the diagonal, since the metric is obtained
 from the embedding by restriction to the tangent space ``T_p\mathcal M`` at ``p``.
 """
-function local_metric(M::Sphere{Tuple{Int}, ℝ}, p, ::DefaultOrthonormalBasis)
+function local_metric(M::Sphere{ℝ, Tuple{Int}}, p, ::DefaultOrthonormalBasis)
     n = get_parameter(M.size)[1]
     return Diagonal(ones(eltype(p), n))
 end
 function local_metric(
-        ::Sphere{TypeParameter{Tuple{n}}, ℝ},
+        ::Sphere{ℝ, TypeParameter{Tuple{n}}},
         p,
         B::DefaultOrthonormalBasis,
     ) where {n}
@@ -509,17 +509,17 @@ function ManifoldsBase.retract_project_fused!(M::AbstractSphere, q, p, X, t::Num
     return project!(M, q, q)
 end
 
-function Base.show(io::IO, ::Sphere{TypeParameter{Tuple{n}}, 𝔽}) where {n, 𝔽}
+function Base.show(io::IO, ::Sphere{𝔽, TypeParameter{Tuple{n}}}) where {n, 𝔽}
     return print(io, "Sphere($(n), $(𝔽))")
 end
-function Base.show(io::IO, M::Sphere{Tuple{Int}, 𝔽}) where {𝔽}
+function Base.show(io::IO, M::Sphere{𝔽, Tuple{Int}}) where {𝔽}
     n = get_parameter(M.size)[1]
     return print(io, "Sphere($(n), $(𝔽); parameter=:field)")
 end
-function Base.show(io::IO, ::ArraySphere{TypeParameter{tn}, 𝔽}) where {tn, 𝔽}
+function Base.show(io::IO, ::ArraySphere{𝔽, TypeParameter{tn}}) where {tn, 𝔽}
     return print(io, "ArraySphere($(join(tn.parameters, ", ")); field=$(𝔽))")
 end
-function Base.show(io::IO, M::ArraySphere{<:Tuple, 𝔽}) where {𝔽}
+function Base.show(io::IO, M::ArraySphere{𝔽, <:Tuple}) where {𝔽}
     n = M.size
     return print(io, "ArraySphere($(join(n, ", ")); field=$(𝔽), parameter=:field)")
 end
@@ -649,7 +649,7 @@ point (1, 0, ..., 0) (called `:south`).
 """
 struct StereographicAtlas <: AbstractAtlas{ℝ} end
 
-function get_chart_index(::Sphere{<:Any, ℝ}, ::StereographicAtlas, p)
+function get_chart_index(::Sphere{ℝ}, ::StereographicAtlas, p)
     if p[1] < 0
         return :south
     else
@@ -657,7 +657,7 @@ function get_chart_index(::Sphere{<:Any, ℝ}, ::StereographicAtlas, p)
     end
 end
 
-function get_parameters!(::Sphere{<:Any, ℝ}, x, ::StereographicAtlas, i::Symbol, p)
+function get_parameters!(::Sphere{ℝ}, x, ::StereographicAtlas, i::Symbol, p)
     if i === :north
         return x .= p[2:end] ./ (1 + p[1])
     else
@@ -665,7 +665,7 @@ function get_parameters!(::Sphere{<:Any, ℝ}, x, ::StereographicAtlas, i::Symbo
     end
 end
 
-function get_point!(::Sphere{<:Any, ℝ}, p, ::StereographicAtlas, i::Symbol, x)
+function get_point!(::Sphere{ℝ}, p, ::StereographicAtlas, i::Symbol, x)
     xnorm2 = dot(x, x)
     if i === :north
         p[1] = (1 - xnorm2) / (xnorm2 + 1)
@@ -677,7 +677,7 @@ function get_point!(::Sphere{<:Any, ℝ}, p, ::StereographicAtlas, i::Symbol, x)
 end
 
 function get_coordinates_induced_basis!(
-        M::Sphere{<:Any, ℝ},
+        M::Sphere{ℝ},
         Y,
         p,
         X,
@@ -697,7 +697,7 @@ function get_coordinates_induced_basis!(
 end
 
 function get_vector_induced_basis!(
-        M::Sphere{<:Any, ℝ},
+        M::Sphere{ℝ},
         Y,
         p,
         X,
@@ -728,7 +728,7 @@ function get_vector_induced_basis!(
 end
 
 function local_metric(
-        M::Sphere{<:Any, ℝ},
+        M::Sphere{ℝ},
         p,
         B::InducedBasis{ℝ, TangentSpaceType, StereographicAtlas, Symbol},
     )