+++ test/manifolds/sphere_symmetric_matrices.jl
@@ -15,7 +15,7 @@ include("../header.jl")
         @test representation_size(M) == (3, 3)
         @test base_manifold(M) === M
         @test !is_flat(M)
-        @test typeof(get_embedding(M)) === ArraySphere{TypeParameter{Tuple{3, 3}}, ℝ}
+        @test typeof(get_embedding(M)) === ArraySphere{ℝ, TypeParameter{Tuple{3, 3}}}
         @test check_point(M, A) === nothing
         @test_throws ManifoldDomainError is_point(M, B; error = :error)
         @test_throws ManifoldDomainError is_point(M, C; error = :error)
@@ -77,6 +77,6 @@ include("../header.jl")
     @testset "field parameter" begin
         M = SphereSymmetricMatrices(3; parameter = :field)
         @test repr(M) == "SphereSymmetricMatrices(3, ℝ; parameter=:field)"
-        @test typeof(get_embedding(M)) === ArraySphere{Tuple{Int, Int}, ℝ}
+        @test typeof(get_embedding(M)) === ArraySphere{ℝ, Tuple{Int, Int}}
     end
 end